# apps/api/src/issues/issue-routes.ts
import type { FastifyInstance } from 'fastify';
import type { AppConfig } from '../config/env.js';
import { giteaFor } from '../gitea/request.js';
import { getIssue, searchIssues } from './issue-service.js';

export async function registerIssueRoutes(app: FastifyInstance, config: AppConfig): Promise<void> {
  app.get('/api/repositories', async (request) => giteaFor(request, config.giteaBaseUrl).repositories());
  app.get('/api/issues', async (request) => {
    const query = request.query as Record<string, string | undefined>;
    return searchIssues(giteaFor(request, config.giteaBaseUrl), {
      q: query.q,
      repository: query.repository,
      state: query.state as 'open' | 'closed' | 'all' | undefined,
      assignee: query.assignee,
      milestone: query.milestone,
      labels: query.label?.split(',').filter(Boolean),
      page: Number(query.page ?? 1),
      limit: Number(query.limit ?? 50),
    });
  });
  app.get('/api/issues/:owner/:repo/:number', async (request) => {
    const params = request.params as { owner: string; repo: string; number: string };
    return getIssue(giteaFor(request, config.giteaBaseUrl), { owner: params.owner, name: params.repo }, Number(params.number));
  });
  app.post('/api/repositories/:owner/:repo/issues', async (request, reply) => {
    const params = request.params as { owner: string; repo: string };
    const issue = await giteaFor(request, config.giteaBaseUrl).createIssue({ owner: params.owner, name: params.repo }, request.body);
    return reply.code(201).send(issue);
  });
  app.patch('/api/issues/:owner/:repo/:number', async (request) => {
    const params = request.params as { owner: string; repo: string; number: string };
    return giteaFor(request, config.giteaBaseUrl).updateIssue({ owner: params.owner, name: params.repo }, Number(params.number), request.body);
  });
  app.get('/api/issues/:owner/:repo/:number/comments', async (request) => {
    const params = request.params as { owner: string; repo: string; number: string };
    return giteaFor(request, config.giteaBaseUrl).comments({ owner: params.owner, name: params.repo }, Number(params.number));
  });
  app.post('/api/issues/:owner/:repo/:number/comments', async (request, reply) => {
    const params = request.params as { owner: string; repo: string; number: string };
    const body = (request.body as { body?: string }).body?.trim();
    if (!body) return reply.code(422).send({ error: 'Comment body is required' });
    const comment = await giteaFor(request, config.giteaBaseUrl).createComment({ owner: params.owner, name: params.repo }, Number(params.number), body);
    return reply.code(201).send(comment);
  });
}